# sql/catalog/resolver: remove an allocation, add a benchmark